+++ auth-service/routes/auth/get-profile.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect, beforeEach, vi, afterEach } from "vitest";
 import request from "supertest";
 import express from "express";
-import { createApp } from "../../app.ts";
+import { createApp } from "../../http.ts";
 import { getCsrfToken } from "./_test-helpers.ts";
 
 // Mock the email package
@@ -11,7 +11,7 @@ describe("Get Profile Route", () => {
   let app: express.Express;
 
   beforeEach(() => {
-    app = createApp({ callbacks: {}});
+    app = createApp({ callbacks: {} });
   });
 
   afterEach(() => {
