+++ auth-service/routes/auth/logout.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect, vi, beforeEach } from "vitest";
 import request from "supertest";
 import express from "express";
-import { createApp } from "../../app.ts";
+import { createApp } from "../../http.ts";
 import { getCsrfToken } from "./_test-helpers.ts";
 
 // Mock the email package
