+++ auth-service/routes/auth/resend-verification.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect, vi, beforeEach } from "vitest";
 import request from "supertest";
 import express from "express";
-import { createApp } from "../../app.ts";
+import { createApp } from "../../http.ts";
 import passport from "passport";
 import { testRateLimiting } from "./_test-helpers.ts";
 import { EmailClient } from "@saflib/email";
@@ -23,7 +23,7 @@ describe("Resend Verification Route", () => {
     vi.clearAllMocks();
     (passport as any)._serializers = [];
     (passport as any)._deserializers = [];
-    app = createApp({ callbacks: {}});
+    app = createApp({ callbacks: {} });
     vi.spyOn(EmailClient.prototype, "sendEmail");
   });
