+++ auth-service/routes/auth/update-profile.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect, beforeEach, vi, afterEach } from "vitest";
 import request from "supertest";
 import express from "express";
-import { createApp } from "../../app.ts";
+import { createApp } from "../../http.ts";
 import { getCsrfToken } from "./_test-helpers.ts";
 
 // Mock the email package
@@ -19,7 +19,7 @@ describe("Update Profile Route", () => {
   let app: express.Express;
 
   beforeEach(() => {
-    app = createApp({ callbacks: {}});
+    app = createApp({ callbacks: {} });
   });
 
   afterEach(() => {
